Make sure wait() isn't called on the event loop
ClientRequest.end() makes the actual HTTP request, the response for which is received over the HTTPClientHandler.
Further, the HTTPClientHandler may either execute HTTP redirection i.e creating a new ClientRequest or simply
calling the `callback` supplied by the user. In both the cases we could again call ClientRequest.end() on the event loop.
This isn't good because this method internally invokes `ClientBootstrap.connect(...).wait()` and swift-nio strictly disallows
calling `wait()` on the event loop. This fix moves the redirection invocations and the callback invocations to a DispatchQueue.
On the DispatchQueue, whenever we wish to work with the `Channel`, we call back into the event loop.