tests: introduce a compiler flag to enable ephemeral ports
Owing to a limitation in ClientRequest, the use of ephemeral ports needs configuration of the ephemeral port range. This configuration is done using the `sysctl` command on both Linux and macOS. While sysctl appears to be working without sudo permissions on Linux, it needs sudo on macOS. This forces `swift test` to stop at the password prompt. This behaviour change needs to be avoided. This commit makes the use of ephemeral ports optional by introducing a compiler flag called `USE_EPHEMERAL_PORTS`.

To run Kitura-NIO tests with ephemeral ports issue: `swift test -Xswiftc -DUSE_EPHEMERAL_PORTS

On Travis CI, we could continue using ephemeral ports on both macOS and Linux. For this, the `.swift-test-macOS` and `.swift-test-linux` are added.