tests: Use ephemeral ports by default
Instead of restricting the ephemeral port range to accommodate port numbers in an Int16 (a known limitation of `ClientRequest`), we can cast Int16s to UInt16s and vice versa. This removes the need to have the USE_EPHEMERAL_PORTS compiler flag and the script that issues commands to restrict the ephemeral port range.