HTTPServerResponse can hold a strong Channel ref
HTTPServerResponse currently has a weak reference to the Channel over which the response will be written. Since a retain cycle between Channel and HTTPServerResponse isn't likely, it makes sense to keep a strong reference to the Channel. This helps with an edge case seen with TechEmpower where the Channel gets 'prematurely' cleaned-up before the last response could be written on a longstanding connection. Keeping a strong reference also eliminates some nil checks, makind the code easier to read.