Don't log errors when a closed channel is detected
In HTTPServerResponse, `channel` and `handler` are weak references to the
underlying NIO channel and the channel handler. These weak vars are used
as a signaling mechanism to indicate a closed connection. While writing
responses back to the underlying channel, we might detect both of this as
nils, which simply means they were deallocated because, most likely, the
underlying connection was closed. This condition does not qualify to be
an error condition.