Pipelining tests: sort responses before comparison
A lot of intermittent failures are being seen with the pipelining tests on Travis. The failures are not reproducible locally. This commit tries to reduce the likelihood of failures by sorting the received responses before comparing them with the expected responses. This is to take care of cases where we might receive responses in an order different from the requests made.