Support WebSocket URIs with query parameters
On receiving a WebSocket upgrade request, the WebSocketUpgrader first checks if
a WebSocket service has been registered at the requested URI. If not, the
upgrade fails. It may so happen that for a WebSocket service registered at a
URL like `/ws`, a WebSocket client might send related query params, example,
`/ws?p1=q1&&p2=q2`. To check if a service is registered at such a URI, it is
essential to strip off the query params.