Add default Accept header and port number to Host (#213)
* Add default Accept header and port number to Host

Kitura-net adds a default Accept header and appends port number to host name if
non-default ports(443/80) are used. In order to maintain parity default Accept
header is added and port name is appended if non default port is used.

* Add default Accept header and port number to Host

Kitura-net adds a default Accept header and appends port number to host name if
non-default ports(443/80) are used. In order to maintain parity default Accept
header is added and port name is appended if non default port is used.