Gracefully close active connections on an HTTPServer stop
If the HTTPServer is stopped while some connections are active, the connections
need to be closed in the `stop()` method. Currently, the stop() method only
closes the listening channel. The child channels, representing the active
connections, also need to be closed. The right way to do this is to allow the
ongoing transfers to complete and gracefully close these channels. To implement
this, we have borrowed the `ServerQuiescingHelper` code from `NIOExtras`. On a
server stop event, the server channel is closed. The open child channels that
represent active connections are then collected, they are allowed to serve
outstanding requests and then shutdown once they are all done. This also
ensures a graceful connection shutdown on the clients.

Initialize the quiescing helper in `listen()`

An HTTPServer.stop() does not destroy the server. It only stops it.
The server can be restarted by calling listen(). This means the
right place to initialize a quiescing helper is in the beginning
of the listen() implementation.